Fixed StackImplementation error

+++ Moderate/StackImplementation.java
@@ -40,8 +40,8 @@ public class StackImplementation {
         size--; 
         
         if (((double)size / storage.length) <= 0.25) {
-		double[] temp = storage; 
-		storage = new double[storage.length / 2]; 
+		int[] temp = storage; 
+		storage = new int[storage.length / 2]; 
 		for (int i = 0; i < size; i++) {
 			storage[i] = temp[i]; 
 		}
